added serverNamesMatch() to ui/lib/toolUtils
